Initial commit
Working on the documentation of diogenes (diogenes.org) is still in
progress. I will now begin to change the code of the Diogenes modules.

First alpha completed. Now working on concurrency and the
implementation of the verba continentia search. An aside is the work
on the dumper for other types of searches (grammatically parsed
searches, etc.)